Module Octez_rpc_process.Directory

type applied_watcher_kind =

| Empty

| Filled of (Tezos_store.Store.chain_store * Tezos_store.Store.Block.t) Lwt_watcher.input

val build_rpc_directory : Tezos_version.Octez_node_version.t -> Octez_node_config.Config_file.t -> Tezos_store.Store.t option Stdlib.ref -> head_watcher: (Tezos_base.TzPervasives.Block_hash.t * Tezos_base.Block_header.t) Lwt_watcher.input -> applied_blocks_watcher:applied_watcher_kind Stdlib.ref -> unit Tezos_rpc.Directory.t

build_rpc_directory node_version config dynamic_store ~head_watcher ~applied_blocks_watcher builds the Tezos RPC directory for the rpc process. RPCs handled here are not forwarded to the node.

head_watcher is the wrapped monitor_head stream on which the clients will listen to.

applied_blocks_watcher is similar to head_watcher but for the applied_blocks RPC.