Module Octez_rpc_process.Forward_handler

val callback : acl:Tezos_rpc_http_server.RPC_server.Acl.t -> Tezos_rpc_http_server.RPC_server.server -> Tezos_rpc_http_server.RPC_middleware.forwarder_resources -> string -> Tezos_rpc_http_server.RPC_server.callback

callback server forwarder_resources socket_path redirects all the traffic received by the given server to the socket determined by socket_path if it fails to be resolved locally. The forwarding happens only if the acl rules are satisfied.

val socket_forwarding_uri : string

val build_socket_redirection_ctx : string -> Cohttp_lwt_unix.Net.ctx